Move BuildResult out of legacy sync
Bug: n/a
Test: n/a

AOSP: 175db03df86fe2eeec561943d5ea57935427fca5
