+++ base/src/com/google/idea/blaze/base/buildview/BazelExecService.kt
@@ -4,12 +4,12 @@ import com.google.devtools.build.lib.buildeventstream.BuildEventStreamProtos.Bui
 import com.google.idea.blaze.base.buildview.events.BuildEventParser
 import com.google.idea.blaze.base.command.BlazeCommand
 import com.google.idea.blaze.base.command.BlazeCommandName
+import com.google.idea.blaze.base.command.buildresult.BuildResult
 import com.google.idea.blaze.base.command.buildresult.BuildResultHelperBep
 import com.google.idea.blaze.base.command.buildresult.BuildResultParser
 import com.google.idea.blaze.base.model.primitives.WorkspaceRoot
 import com.google.idea.blaze.base.scope.BlazeContext
 import com.google.idea.blaze.base.sync.aspects.BlazeBuildOutputs
-import com.google.idea.blaze.base.sync.aspects.BuildResult
 import com.google.idea.blaze.common.Interners
 import com.google.idea.blaze.common.PrintOutput
 import com.google.protobuf.CodedInputStream
