+++ base/src/com/google/idea/blaze/base/command/buildresult/BuildResult.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.idea.blaze.base.sync.aspects;
+package com.google.idea.blaze.base.command.buildresult;
 
 import java.util.Objects;
