+++ base/src/com/google/idea/blaze/base/logging/utils/BuildPhaseSyncStats.java
@@ -20,7 +20,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.idea.blaze.base.model.primitives.TargetExpression;
 import com.google.idea.blaze.base.scope.scopes.TimingScopeListener.TimedEvent;
 import com.google.idea.blaze.base.settings.BuildBinaryType;
-import com.google.idea.blaze.base.sync.aspects.BuildResult;
+import com.google.idea.blaze.base.command.buildresult.BuildResult;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
