+++ base/src/com/google/idea/blaze/base/run/BlazeBeforeRunCommandHelper.java
@@ -42,7 +42,7 @@ import com.google.idea.blaze.base.scope.scopes.ProblemsViewScope;
 import com.google.idea.blaze.base.scope.scopes.ToolWindowScope;
 import com.google.idea.blaze.base.settings.Blaze;
 import com.google.idea.blaze.base.settings.BlazeUserSettings;
-import com.google.idea.blaze.base.sync.aspects.BuildResult;
+import com.google.idea.blaze.base.command.buildresult.BuildResult;
 import com.google.idea.blaze.base.toolwindow.Task;
 import com.intellij.openapi.project.Project;
 import java.io.IOException;
